motoblog/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-taggit==3.0.0
Markdown==3.4.1
psycopg2==2.9.5
sqlparse==0.4.4
tzdata==2023.4
